requirements.txt
# ============================================
# J.A. Uniforms Pricing Tool - Dependencies
# ============================================

# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Werkzeug==3.0.0

# Database & Migrations
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Email
Flask-Mail==0.9.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Production Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
Flask-Limiter==3.5.0
pytz==2023.3

# ============================================
# Testing Dependencies (install with: pip install -r requirements.txt)
# ============================================
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2